Add support for installing an additional target
When the rustup buildpack runs, it installs the specified Rust tools. It is possible for rustup to install additional toolsets. This is helpful when cross-compiling. For buildpacks, this would be most useful to build a binary that runs on the Paketo tiny stack.

If `$BP_RUST_TARGET` is set, the target in it will be installed after the Rust toolchain is installed. This amounts to running `rustup target add`. Valid targets for this option are the same as that command.

If `$BP_RUST_TARGET` is not set, the buildpack will install an additional target when run on the Paketo Tiny stack. This stack requires using musl instead of gnu libc, so the buildpack will automatcally set it.

Resolves #44.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>