Fixes layer cache reloading problem
In libpak 1.60.0, a protection was added where if a layer has the build or cache flag set and it is empty, the buildpack will reload it. This prevents weirdness when the layer has not been properly restored. This protection was being triggered for the Rust layer because the layer is build + cache, but doesn't contain any files (intentionally).

We want it to be build + cache because we use the metadata to determine if the rustup operation should run to reinstall the layer, however, the actual files that get installed go into RUSTUP_HOME, which is a different layer. This is just how rustup works and we can't separate the two. At the same time, stuffing everything into a single layer would be complicated and huge, which is why we've made this two separate layers.

The net result is that to restore proper caching behavior we need to write at least one file to the layer. This PR writes an empty 'marker' file. The file does nothing except ensure that libpack's layer reloading only runs if the layer actually had a problem being restored.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>