Fixes update bug with rustup-init
The dependency id was not correct in pipeline-descriptor, thus updates were being detected but not reflected into buildpack.toml. This change syncs up pipeline-descriptor.yml with the id names in buildpack.toml.

After this completes, we need to run the two update jobs as there are pending updates we are missing.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>